requirements.txt
Django==1.11.3
django-wysiwyg-redactor==0.5.1
olefile==0.44
Pillow==4.2.1
pytz==2017.2
